build: reduce cache size, extract ISO
By using `file.cached` instead of `file.managed`, only a single copy of
the ISO is created.
The extraction of the ISO uses the cached file directly from the cache
location, determined by `cp.is_cached`.
For extracting the ISO without having to mount it as loop-back device,
`bsdtar` is required which is part of the `libarchive-tools` pkg on
Debian/Ubuntu. In case this pkg isn't installed yet, the state needs to
be executed as `root`, as required by the pkg manager.